# swagger.json
{"openapi":"3.0.1","info":{"title":"Wedoogift API","description":"Wedoogift api information","version":"0.0.1"},"servers":[{"url":"http://localhost:8001","description":"Generated server url"}],"paths":{"/api/user":{"get":{"tags":["user-controller"],"summary":"Get a user by its id","operationId":"getUser","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"put":{"tags":["user-controller"],"summary":"Update an user","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"post":{"tags":["user-controller"],"summary":"Create a new user","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["user-controller"],"summary":"Delete an user","operationId":"deleteUser","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]}}}}}}},"/api/payment":{"post":{"tags":["payment-controller"],"summary":"Add a payment","operationId":"addPayment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentDto"}}}}}}},"/api/deposit":{"post":{"tags":["deposit-controller"],"summary":"Add a deposit","operationId":"addDeposit","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepositDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DepositDto"}}}}}}},"/api/company":{"post":{"tags":["company-controller"],"summary":"Create a new company","operationId":"createUser_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CompanyDto"}}}}}}}},"components":{"schemas":{"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"balance":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"}}},"PaymentDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"amount":{"type":"number"},"receiver":{"type":"string"},"object":{"type":"string"},"userId":{"type":"integer","format":"int32"}}},"DepositDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"amount":{"type":"number"},"beginDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"},"companyId":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"type":{"type":"string","enum":["GIFT","MEAL"]}}},"CompanyDto":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"publicName":{"type":"string"},"corporateName":{"type":"string"},"siren":{"type":"string"},"siret":{"type":"string"},"balance":{"type":"number"}}}}}}